requirements.txt
discord.py==2.4.0
httpx==0.27.2
tqdm==4.66.5
python-dotenv==1.0.1

